requirement.txt
# packages in environment at /opt/anaconda3/envs/py3_tf1:
#
# Name                    Version                   Build  Channel
absl-py                   0.13.0                   pypi_0    pypi
astor                     0.8.1                    pypi_0    pypi
bert-base                 0.0.9                    pypi_0    pypi
ca-certificates           2021.7.5             hecd8cb5_1  
cached-property           1.5.2                    pypi_0    pypi
certifi                   2021.5.30        py36hecd8cb5_0  
dataclasses               0.8                      pypi_0    pypi
gast                      0.5.0                    pypi_0    pypi
gensim                    4.0.1                    pypi_0    pypi
google-pasta              0.2.0                    pypi_0    pypi
gputil                    1.4.0                    pypi_0    pypi
grpcio                    1.36.1                   pypi_0    pypi
h5py                      3.1.0                    pypi_0    pypi
importlib-metadata        4.6.1                    pypi_0    pypi
joblib                    1.0.1                    pypi_0    pypi
keras-applications        1.0.8                    pypi_0    pypi
keras-preprocessing       1.1.2                    pypi_0    pypi
libcxx                    10.0.0                        1  
libffi                    3.3                  hb1e8313_2  
markdown                  3.3.4                    pypi_0    pypi
ncurses                   6.2                  h0a44026_1  
numpy                     1.19.5                   pypi_0    pypi
openssl                   1.1.1k               h9ed2024_0  
pip                       21.1.3           py36hecd8cb5_0  
protobuf                  3.17.3                   pypi_0    pypi
python                    3.6.13               h88f2d9e_0  
pyzmq                     22.1.0                   pypi_0    pypi
readline                  8.1                  h9ed2024_0  
scikit-learn              0.24.2                   pypi_0    pypi
scipy                     1.5.4                    pypi_0    pypi
seqeval                   1.2.2                    pypi_0    pypi
setuptools                52.0.0           py36hecd8cb5_0  
six                       1.16.0                   pypi_0    pypi
smart-open                5.1.0                    pypi_0    pypi
sqlite                    3.36.0               hce871da_0  
tensorboard               1.14.0                   pypi_0    pypi
tensorflow                1.14.0                   pypi_0    pypi
tensorflow-estimator      1.14.0                   pypi_0    pypi
tensorflow-serving-api    1.14.0                   pypi_0    pypi
termcolor                 1.1.0                    pypi_0    pypi
threadpoolctl             2.2.0                    pypi_0    pypi
tk                        8.6.10               hb0a8c7a_0  
typing-extensions         3.10.0.0                 pypi_0    pypi
werkzeug                  2.0.1                    pypi_0    pypi
wheel                     0.36.2             pyhd3eb1b0_0  
wrapt                     1.12.1                   pypi_0    pypi
xz                        5.2.5                h1de35cc_0  
zipp                      3.5.0                    pypi_0    pypi
zlib                      1.2.11               h1de35cc_3  
